Add files via upload
The footboard avoider project is a safety system designed to prevent accidents caused by passengers standing on the footboard of vehicles. It uses an IR sensor to detect the presence of a person on the footboard. When someone is detected, the system triggers a buzzer and LED to alert the driver and passengers, ensuring immediate action. A relay module and a regulated power supply manage the system's operations. This project provides a cost-effective and efficient solution to enhance passenger safety and reduce the risk of accidents.